4th commit : Used H2 in-memory database

+++ src/main/java/org/example/demoapplication/dto/response/StudentResponse.java
@@ -7,20 +7,20 @@ import jakarta.validation.constraints.NotBlank;
 public class StudentResponse {
     @Min(1)
     @NotBlank
-    private int id;
+    private Long id;
     @NotBlank
     private String name;
     @Min(value = 1,message = "Age can not be less than 1")
     @Max(value = 150,message = "Age can not be greater than 150")
     private int age;
 
-    public StudentResponse(int id,String name, int age){
+    public StudentResponse(Long id,String name, int age){
         this.id = id;
         this.name = name;
         this.age = age;
     }
 
-    public int getID(){return id;}
+    public Long getID(){return id;}
 
     public String getName(){
         return name;
